+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v4.19.0
+  - The formatter `formatC` accepts a third parameter to forces the target currency. It overwrites the global options `currencyTarget`.
+  - If the global option `currencySource` is undefined, no conversion is done when using the formatters `convCurr` and `formatC`.
+
 ## v4.18.0
   - Release February 14th 2024
   - [EE] Support dynamic pictures into Shapes for DOCX template only. Create images with rounded corners, or images with special shapes.